parsing scale from json, changing default scale to 255, fixing scenes under test, adding commented code for a no fog mode

+++ RayApi.hpp
@@ -20,7 +20,7 @@ void render(void);
 // void doDebugRender(void);
 // uint32_t get4(void);
 // uint32_t getNext(void);
-void setScale(const float s);
+void setScale(const double s);
 // void debug2(void);
 // void setFrameSleep(const int s);
 void onRuntimeInitialized(void);
